fix: final touches to the project

+++ src/modules/property-me/property-me.service.ts
@@ -1,5 +1,5 @@
 import { CACHE_KEYS } from '@/cache/cache-keys';
-import { CacheUtilsService, CollectionDiffService } from '@/common/services';
+import { CacheUtilsService, CollectionDiffService, RedisService } from '@/common/services';
 import { CreateImagePropertyMeDto, CreateImagesPropertyMeDto, CreatePropertyMeDto, PaginationPropertyMeDto, UpdatePropertyMeDto } from './dto';
 import { Injectable } from '@nestjs/common';
 import { IMAGE_TYPE, Location, Property, User } from 'generated/prisma';
@@ -8,6 +8,7 @@ import { PropertyService, TransactionService, ImagesPropertyService } from './se
 import { NormalizerService } from './services/normalizer';
 import { LocationService } from '@/modules/location/location.service';
 import { ImagesService } from '@/modules/images/images.service';
+import { TTL } from '@/cache/ttls';
 ;
 
 @Injectable()
@@ -21,19 +22,25 @@ export class PropertyMeService {
     private readonly transactionService: TransactionService,
     private readonly normalizerService: NormalizerService,
     private readonly locationService: LocationService,
-    private readonly imagesService: ImagesService
+    private readonly imagesService: ImagesService,
+    private readonly redisService: RedisService
   ) { }
 
-  async location(departmentId: Location['departmentId'], provinceId: Location['provinceId'], districtId: Location['districtId']){
+  async location(departmentId: Location['departmentId'], provinceId: Location['provinceId'], districtId: Location['districtId']) {
     return await this.locationService.findLocation(departmentId, provinceId, districtId);
   }
 
   async create(dto: CreatePropertyMeDto, userId: User['id']) {
     const location = await this.location(dto.departmentId, dto.provinceId, dto.districtId);
-    await this.propertyService.create({...dto}, location.id, userId);
-    this.cacheUtilsService.deleteKeys(CACHE_KEYS.PROPERTIES_ME);
+    await Promise.all([
+      this.propertyService.create({ ...dto }, location.id, userId),
+      this.cacheUtilsService.deleteKeys([
+        `${CACHE_KEYS.PROPERTIES_ME}/${userId}`,
+        CACHE_KEYS.PROPERTY_PUBLIC
+      ]),
+    ])
     return {
-      message: 'Property created successfully',
+      message: 'Propiedad creada correctamente',
     }
   }
 
@@ -42,8 +49,8 @@ export class PropertyMeService {
   }
 
   async createImageMain(dto: CreateImagePropertyMeDto) {
-    this.cacheUtilsService.deleteKeys([`${CACHE_KEYS.PROPERTY_IMAGES}/${dto.propertyId}`]);
     const created = await this.imagesPropertyService.createMain(dto);
+    await this.cacheUtilsService.deleteKeys([`${CACHE_KEYS.PROPERTY_IMAGES}/${dto.propertyId}`, CACHE_KEYS.PROPERTY_PUBLIC]);
     return {
       image: created.url
     }
@@ -51,7 +58,7 @@ export class PropertyMeService {
 
   async createImagesGallery(dto: CreateImagesPropertyMeDto) {
     const currentImages = (await this.imagesPropertyService.images(dto.propertyId)).filter(image => image.type === IMAGE_TYPE.GALLERY).map(image => image.url);
-  
+
     const { adds, deletes } = this.collectionDiffService.preparedData(dto.images.map(image => image.url), currentImages);
 
     const addsObjects = dto.images.filter(image => adds.includes(image.url));
@@ -59,38 +66,49 @@ export class PropertyMeService {
       this.imagesPropertyService.deleteImages(deletes),
       this.imagesService.remove(deletes)
     ])
-    this.cacheUtilsService.deleteKeys([`${CACHE_KEYS.PROPERTY_IMAGES}/${dto.propertyId}`]);
+    this.cacheUtilsService.deleteKeys([`${CACHE_KEYS.PROPERTY_IMAGES}/${dto.propertyId}`, CACHE_KEYS.PROPERTY_PUBLIC]);
     const created = await this.imagesPropertyService.createGallery({ ...dto, images: addsObjects });
     return {
       images: created.map(image => image.url)
     }
   }
 
-
   async findAll(queryParams: PaginationPropertyMeDto, userId: User['id'],) {
-    return await this.propertyService.findAll(userId, queryParams);
+    const key = this.cacheUtilsService.generateCacheKey(CACHE_KEYS.PROPERTIES_ME, queryParams, userId);
+    const cached = await this.redisService.get(key);
+    if(cached){
+      return JSON.parse(cached);
+    } else {
+      const data = await this.propertyService.findAll(userId, queryParams);
+      await this.redisService.set(key, JSON.stringify(data), 'EX', TTL.ONE_HOUR);
+      return data
+    }
   }
 
   async findOneWithRelations(id: Property['id'], userId: User['id']) {
     return await this.propertyService.findOneWithRelations(id, userId);
   }
 
-  async update(property: PropertyFormatted, updateProperty: UpdatePropertyMeDto) {
+  async update(property: PropertyFormatted, updateProperty: UpdatePropertyMeDto, userId: User['id']) {
     const location = await this.location(updateProperty.departmentId, updateProperty.provinceId, updateProperty.districtId);
     const propertyNormalized = this.normalizerService.normalizeProperty(updateProperty);
     const { adds, deletes } = this.collectionDiffService.preparedData(propertyNormalized.servicesId, property.servicesId);
-    await this.transactionService.update(deletes, location.id, { ...propertyNormalized, servicesId: adds }, String(property.id));
-    this.cacheUtilsService.deleteKeys([CACHE_KEYS.PROPERTIES_ME, `${CACHE_KEYS.PROPERTY_ME}/${property.id}`, `${CACHE_KEYS.PROPERTY_DETAIL}/${property.id}`]);
+    await Promise.all([
+      this.transactionService.update(deletes, location.id, { ...propertyNormalized, servicesId: adds }, String(property.id)),
+      this.cacheUtilsService.deleteKeys([`${CACHE_KEYS.PROPERTIES_ME}/${userId}`, `${CACHE_KEYS.PROPERTY_ME}/${property.id}`, `${CACHE_KEYS.PROPERTY_DETAIL}/${property.id}`, CACHE_KEYS.PROPERTY_PUBLIC]),
+    ])
     return {
-      message: 'Property updated successfully',
+      message: 'Propiedad actualizada correctamente',
     }
   }
 
-  async changeStatus(property: PropertyFormatted) {
-    await this.propertyService.changeStatus(String(property.id), property.availability);
-    this.cacheUtilsService.deleteKeys([CACHE_KEYS.PROPERTIES_ME, `${CACHE_KEYS.PROPERTY_ME}/${property.id}`, `${CACHE_KEYS.PROPERTY_DETAIL}/${property.id}`]);
+  async changeStatus(property: PropertyFormatted, userId: User['id']) {
+    await Promise.all([
+      this.propertyService.changeStatus(String(property.id), property.availability),
+      this.cacheUtilsService.deleteKeys([`${CACHE_KEYS.PROPERTIES_ME}/${userId}}`, `${CACHE_KEYS.PROPERTY_ME}/${property.id}`, `${CACHE_KEYS.PROPERTY_DETAIL}/${property.id}`, CACHE_KEYS.PROPERTY_PUBLIC]),
+    ])
     return {
-      message: 'Property status changed successfully',
+      message: 'Estado de la propiedad actualizada correctamente',
     }
   }
 }
